8319522: Consolidate heap dump writer and other utilities
