+++ src/hotspot/share/services/heapDumper.cpp
@@ -58,7 +58,7 @@
 #include "runtime/vmThread.hpp"
 #include "runtime/timerTrace.hpp"
 #include "services/heapDumper.hpp"
-#include "services/heapDumperCompression.hpp"
+#include "services/heapDumperWriter.hpp"
 #include "services/threadService.hpp"
 #include "utilities/checkedCast.hpp"
 #include "utilities/macros.hpp"
@@ -67,660 +67,11 @@
 #include "os_linux.hpp"
 #endif
 
-/*
- * HPROF binary format - description copied from:
- *   src/share/demo/jvmti/hprof/hprof_io.c
- *
- *
- *  header    "JAVA PROFILE 1.0.2" (0-terminated)
- *
- *  u4        size of identifiers. Identifiers are used to represent
- *            UTF8 strings, objects, stack traces, etc. They usually
- *            have the same size as host pointers.
- * u4         high word
- * u4         low word    number of milliseconds since 0:00 GMT, 1/1/70
- * [record]*  a sequence of records.
- *
- *
- * Record format:
- *
- * u1         a TAG denoting the type of the record
- * u4         number of *microseconds* since the time stamp in the
- *            header. (wraps around in a little more than an hour)
- * u4         number of bytes *remaining* in the record. Note that
- *            this number excludes the tag and the length field itself.
- * [u1]*      BODY of the record (a sequence of bytes)
- *
- *
- * The following TAGs are supported:
- *
- * TAG           BODY       notes
- *----------------------------------------------------------
- * HPROF_UTF8               a UTF8-encoded name
- *
- *               id         name ID
- *               [u1]*      UTF8 characters (no trailing zero)
- *
- * HPROF_LOAD_CLASS         a newly loaded class
- *
- *                u4        class serial number (> 0)
- *                id        class object ID
- *                u4        stack trace serial number
- *                id        class name ID
- *
- * HPROF_UNLOAD_CLASS       an unloading class
- *
- *                u4        class serial_number
- *
- * HPROF_FRAME              a Java stack frame
- *
- *                id        stack frame ID
- *                id        method name ID
- *                id        method signature ID
- *                id        source file name ID
- *                u4        class serial number
- *                i4        line number. >0: normal
- *                                       -1: unknown
- *                                       -2: compiled method
- *                                       -3: native method
- *
- * HPROF_TRACE              a Java stack trace
- *
- *               u4         stack trace serial number
- *               u4         thread serial number
- *               u4         number of frames
- *               [id]*      stack frame IDs
- *
- *
- * HPROF_ALLOC_SITES        a set of heap allocation sites, obtained after GC
- *
- *               u2         flags 0x0001: incremental vs. complete
- *                                0x0002: sorted by allocation vs. live
- *                                0x0004: whether to force a GC
- *               u4         cutoff ratio
- *               u4         total live bytes
- *               u4         total live instances
- *               u8         total bytes allocated
- *               u8         total instances allocated
- *               u4         number of sites that follow
- *               [u1        is_array: 0:  normal object
- *                                    2:  object array
- *                                    4:  boolean array
- *                                    5:  char array
- *                                    6:  float array
- *                                    7:  double array
- *                                    8:  byte array
- *                                    9:  short array
- *                                    10: int array
- *                                    11: long array
- *                u4        class serial number (may be zero during startup)
- *                u4        stack trace serial number
- *                u4        number of bytes alive
- *                u4        number of instances alive
- *                u4        number of bytes allocated
- *                u4]*      number of instance allocated
- *
- * HPROF_START_THREAD       a newly started thread.
- *
- *               u4         thread serial number (> 0)
- *               id         thread object ID
- *               u4         stack trace serial number
- *               id         thread name ID
- *               id         thread group name ID
- *               id         thread group parent name ID
- *
- * HPROF_END_THREAD         a terminating thread.
- *
- *               u4         thread serial number
- *
- * HPROF_HEAP_SUMMARY       heap summary
- *
- *               u4         total live bytes
- *               u4         total live instances
- *               u8         total bytes allocated
- *               u8         total instances allocated
- *
- * HPROF_HEAP_DUMP          denote a heap dump
- *
- *               [heap dump sub-records]*
- *
- *                          There are four kinds of heap dump sub-records:
- *
- *               u1         sub-record type
- *
- *               HPROF_GC_ROOT_UNKNOWN         unknown root
- *
- *                          id         object ID
- *
- *               HPROF_GC_ROOT_THREAD_OBJ      thread object
- *
- *                          id         thread object ID  (may be 0 for a
- *                                     thread newly attached through JNI)
- *                          u4         thread sequence number
- *                          u4         stack trace sequence number
- *
- *               HPROF_GC_ROOT_JNI_GLOBAL      JNI global ref root
- *
- *                          id         object ID
- *                          id         JNI global ref ID
- *
- *               HPROF_GC_ROOT_JNI_LOCAL       JNI local ref
- *
- *                          id         object ID
- *                          u4         thread serial number
- *                          u4         frame # in stack trace (-1 for empty)
- *
- *               HPROF_GC_ROOT_JAVA_FRAME      Java stack frame
- *
- *                          id         object ID
- *                          u4         thread serial number
- *                          u4         frame # in stack trace (-1 for empty)
- *
- *               HPROF_GC_ROOT_NATIVE_STACK    Native stack
- *
- *                          id         object ID
- *                          u4         thread serial number
- *
- *               HPROF_GC_ROOT_STICKY_CLASS    System class
- *
- *                          id         object ID
- *
- *               HPROF_GC_ROOT_THREAD_BLOCK    Reference from thread block
- *
- *                          id         object ID
- *                          u4         thread serial number
- *
- *               HPROF_GC_ROOT_MONITOR_USED    Busy monitor
- *
- *                          id         object ID
- *
- *               HPROF_GC_CLASS_DUMP           dump of a class object
- *
- *                          id         class object ID
- *                          u4         stack trace serial number
- *                          id         super class object ID
- *                          id         class loader object ID
- *                          id         signers object ID
- *                          id         protection domain object ID
- *                          id         reserved
- *                          id         reserved
- *
- *                          u4         instance size (in bytes)
- *
- *                          u2         size of constant pool
- *                          [u2,       constant pool index,
- *                           ty,       type
- *                                     2:  object
- *                                     4:  boolean
- *                                     5:  char
- *                                     6:  float
- *                                     7:  double
- *                                     8:  byte
- *                                     9:  short
- *                                     10: int
- *                                     11: long
- *                           vl]*      and value
- *
- *                          u2         number of static fields
- *                          [id,       static field name,
- *                           ty,       type,
- *                           vl]*      and value
- *
- *                          u2         number of inst. fields (not inc. super)
- *                          [id,       instance field name,
- *                           ty]*      type
- *
- *               HPROF_GC_INSTANCE_DUMP        dump of a normal object
- *
- *                          id         object ID
- *                          u4         stack trace serial number
- *                          id         class object ID
- *                          u4         number of bytes that follow
- *                          [vl]*      instance field values (class, followed
- *                                     by super, super's super ...)
- *
- *               HPROF_GC_OBJ_ARRAY_DUMP       dump of an object array
- *
- *                          id         array object ID
- *                          u4         stack trace serial number
- *                          u4         number of elements
- *                          id         array class ID
- *                          [id]*      elements
- *
- *               HPROF_GC_PRIM_ARRAY_DUMP      dump of a primitive array
- *
- *                          id         array object ID
- *                          u4         stack trace serial number
- *                          u4         number of elements
- *                          u1         element type
- *                                     4:  boolean array
- *                                     5:  char array
- *                                     6:  float array
- *                                     7:  double array
- *                                     8:  byte array
- *                                     9:  short array
- *                                     10: int array
- *                                     11: long array
- *                          [u1]*      elements
- *
- * HPROF_CPU_SAMPLES        a set of sample traces of running threads
- *
- *                u4        total number of samples
- *                u4        # of traces
- *               [u4        # of samples
- *                u4]*      stack trace serial number
- *
- * HPROF_CONTROL_SETTINGS   the settings of on/off switches
- *
- *                u4        0x00000001: alloc traces on/off
- *                          0x00000002: cpu sampling on/off
- *                u2        stack trace depth
- *
- *
- * When the header is "JAVA PROFILE 1.0.2" a heap dump can optionally
- * be generated as a sequence of heap dump segments. This sequence is
- * terminated by an end record. The additional tags allowed by format
- * "JAVA PROFILE 1.0.2" are:
- *
- * HPROF_HEAP_DUMP_SEGMENT  denote a heap dump segment
- *
- *               [heap dump sub-records]*
- *               The same sub-record types allowed by HPROF_HEAP_DUMP
- *
- * HPROF_HEAP_DUMP_END      denotes the end of a heap dump
- *
- */
-
-
-// HPROF tags
-
-enum hprofTag : u1 {
-  // top-level records
-  HPROF_UTF8                    = 0x01,
-  HPROF_LOAD_CLASS              = 0x02,
-  HPROF_UNLOAD_CLASS            = 0x03,
-  HPROF_FRAME                   = 0x04,
-  HPROF_TRACE                   = 0x05,
-  HPROF_ALLOC_SITES             = 0x06,
-  HPROF_HEAP_SUMMARY            = 0x07,
-  HPROF_START_THREAD            = 0x0A,
-  HPROF_END_THREAD              = 0x0B,
-  HPROF_HEAP_DUMP               = 0x0C,
-  HPROF_CPU_SAMPLES             = 0x0D,
-  HPROF_CONTROL_SETTINGS        = 0x0E,
-
-  // 1.0.2 record types
-  HPROF_HEAP_DUMP_SEGMENT       = 0x1C,
-  HPROF_HEAP_DUMP_END           = 0x2C,
-
-  // field types
-  HPROF_ARRAY_OBJECT            = 0x01,
-  HPROF_NORMAL_OBJECT           = 0x02,
-  HPROF_BOOLEAN                 = 0x04,
-  HPROF_CHAR                    = 0x05,
-  HPROF_FLOAT                   = 0x06,
-  HPROF_DOUBLE                  = 0x07,
-  HPROF_BYTE                    = 0x08,
-  HPROF_SHORT                   = 0x09,
-  HPROF_INT                     = 0x0A,
-  HPROF_LONG                    = 0x0B,
-
-  // data-dump sub-records
-  HPROF_GC_ROOT_UNKNOWN         = 0xFF,
-  HPROF_GC_ROOT_JNI_GLOBAL      = 0x01,
-  HPROF_GC_ROOT_JNI_LOCAL       = 0x02,
-  HPROF_GC_ROOT_JAVA_FRAME      = 0x03,
-  HPROF_GC_ROOT_NATIVE_STACK    = 0x04,
-  HPROF_GC_ROOT_STICKY_CLASS    = 0x05,
-  HPROF_GC_ROOT_THREAD_BLOCK    = 0x06,
-  HPROF_GC_ROOT_MONITOR_USED    = 0x07,
-  HPROF_GC_ROOT_THREAD_OBJ      = 0x08,
-  HPROF_GC_CLASS_DUMP           = 0x20,
-  HPROF_GC_INSTANCE_DUMP        = 0x21,
-  HPROF_GC_OBJ_ARRAY_DUMP       = 0x22,
-  HPROF_GC_PRIM_ARRAY_DUMP      = 0x23
-};
-
 // Default stack trace ID (used for dummy HPROF_TRACE record)
 enum {
   STACK_TRACE_ID = 1,
   INITIAL_CLASS_COUNT = 200
 };
-
-// Supports I/O operations for a dump
-// Base class for dump and parallel dump
-class AbstractDumpWriter : public CHeapObj<mtInternal> {
- protected:
-  enum {
-    io_buffer_max_size = 1*M,
-    dump_segment_header_size = 9
-  };
-
-  char* _buffer;    // internal buffer
-  size_t _size;
-  size_t _pos;
-
-  bool _in_dump_segment; // Are we currently in a dump segment?
-  bool _is_huge_sub_record; // Are we writing a sub-record larger than the buffer size?
-  DEBUG_ONLY(size_t _sub_record_left;) // The bytes not written for the current sub-record.
-  DEBUG_ONLY(bool _sub_record_ended;) // True if we have called the end_sub_record().
-
-  char* buffer() const                          { return _buffer; }
-  size_t buffer_size() const                    { return _size; }
-  void set_position(size_t pos)                 { _pos = pos; }
-
-  // Can be called if we have enough room in the buffer.
-  void write_fast(const void* s, size_t len);
-
-  // Returns true if we have enough room in the buffer for 'len' bytes.
-  bool can_write_fast(size_t len);
-
-  void write_address(address a);
-
- public:
-  AbstractDumpWriter() :
-    _buffer(nullptr),
-    _size(io_buffer_max_size),
-    _pos(0),
-    _in_dump_segment(false) { }
-
-  // Total number of bytes written to the disk
-  virtual julong bytes_written() const = 0;
-  // Return non-null if error occurred
-  virtual char const* error() const = 0;
-
-  size_t position() const                       { return _pos; }
-  // writer functions
-  virtual void write_raw(const void* s, size_t len);
-  void write_u1(u1 x);
-  void write_u2(u2 x);
-  void write_u4(u4 x);
-  void write_u8(u8 x);
-  void write_objectID(oop o);
-  void write_rootID(oop* p);
-  void write_symbolID(Symbol* o);
-  void write_classID(Klass* k);
-  void write_id(u4 x);
-
-  // Start a new sub-record. Starts a new heap dump segment if needed.
-  void start_sub_record(u1 tag, u4 len);
-  // Ends the current sub-record.
-  void end_sub_record();
-  // Finishes the current dump segment if not already finished.
-  void finish_dump_segment();
-  // Flush internal buffer to persistent storage
-  virtual void flush() = 0;
-};
-
-void AbstractDumpWriter::write_fast(const void* s, size_t len) {
-  assert(!_in_dump_segment || (_sub_record_left >= len), "sub-record too large");
-  assert(buffer_size() - position() >= len, "Must fit");
-  debug_only(_sub_record_left -= len);
-  memcpy(buffer() + position(), s, len);
-  set_position(position() + len);
-}
-
-bool AbstractDumpWriter::can_write_fast(size_t len) {
-  return buffer_size() - position() >= len;
-}
-
-// write raw bytes
-void AbstractDumpWriter::write_raw(const void* s, size_t len) {
-  assert(!_in_dump_segment || (_sub_record_left >= len), "sub-record too large");
-  debug_only(_sub_record_left -= len);
-
-  // flush buffer to make room.
-  while (len > buffer_size() - position()) {
-    assert(!_in_dump_segment || _is_huge_sub_record,
-           "Cannot overflow in non-huge sub-record.");
-    size_t to_write = buffer_size() - position();
-    memcpy(buffer() + position(), s, to_write);
-    s = (void*) ((char*) s + to_write);
-    len -= to_write;
-    set_position(position() + to_write);
-    flush();
-  }
-
-  memcpy(buffer() + position(), s, len);
-  set_position(position() + len);
-}
-
-// Makes sure we inline the fast write into the write_u* functions. This is a big speedup.
-#define WRITE_KNOWN_TYPE(p, len) do { if (can_write_fast((len))) write_fast((p), (len)); \
-                                      else write_raw((p), (len)); } while (0)
-
-void AbstractDumpWriter::write_u1(u1 x) {
-  WRITE_KNOWN_TYPE(&x, 1);
-}
-
-void AbstractDumpWriter::write_u2(u2 x) {
-  u2 v;
-  Bytes::put_Java_u2((address)&v, x);
-  WRITE_KNOWN_TYPE(&v, 2);
-}
-
-void AbstractDumpWriter::write_u4(u4 x) {
-  u4 v;
-  Bytes::put_Java_u4((address)&v, x);
-  WRITE_KNOWN_TYPE(&v, 4);
-}
-
-void AbstractDumpWriter::write_u8(u8 x) {
-  u8 v;
-  Bytes::put_Java_u8((address)&v, x);
-  WRITE_KNOWN_TYPE(&v, 8);
-}
-
-void AbstractDumpWriter::write_address(address a) {
-#ifdef _LP64
-  write_u8((u8)a);
-#else
-  write_u4((u4)a);
-#endif
-}
-
-void AbstractDumpWriter::write_objectID(oop o) {
-  write_address(cast_from_oop<address>(o));
-}
-
-void AbstractDumpWriter::write_rootID(oop* p) {
-  write_address((address)p);
-}
-
-void AbstractDumpWriter::write_symbolID(Symbol* s) {
-  write_address((address)((uintptr_t)s));
-}
-
-void AbstractDumpWriter::write_id(u4 x) {
-#ifdef _LP64
-  write_u8((u8) x);
-#else
-  write_u4(x);
-#endif
-}
-
-// We use java mirror as the class ID
-void AbstractDumpWriter::write_classID(Klass* k) {
-  write_objectID(k->java_mirror());
-}
-
-void AbstractDumpWriter::finish_dump_segment() {
-  if (_in_dump_segment) {
-    assert(_sub_record_left == 0, "Last sub-record not written completely");
-    assert(_sub_record_ended, "sub-record must have ended");
-
-    // Fix up the dump segment length if we haven't written a huge sub-record last
-    // (in which case the segment length was already set to the correct value initially).
-    if (!_is_huge_sub_record) {
-      assert(position() > dump_segment_header_size, "Dump segment should have some content");
-      Bytes::put_Java_u4((address) (buffer() + 5),
-                         (u4) (position() - dump_segment_header_size));
-    } else {
-      // Finish process huge sub record
-      // Set _is_huge_sub_record to false so the parallel dump writer can flush data to file.
-      _is_huge_sub_record = false;
-    }
-
-    _in_dump_segment = false;
-    flush();
-  }
-}
-
-void AbstractDumpWriter::start_sub_record(u1 tag, u4 len) {
-  if (!_in_dump_segment) {
-    if (position() > 0) {
-      flush();
-    }
-
-    assert(position() == 0 && buffer_size() > dump_segment_header_size, "Must be at the start");
-
-    write_u1(HPROF_HEAP_DUMP_SEGMENT);
-    write_u4(0); // timestamp
-    // Will be fixed up later if we add more sub-records.  If this is a huge sub-record,
-    // this is already the correct length, since we don't add more sub-records.
-    write_u4(len);
-    assert(Bytes::get_Java_u4((address)(buffer() + 5)) == len, "Inconsistent size!");
-    _in_dump_segment = true;
-    _is_huge_sub_record = len > buffer_size() - dump_segment_header_size;
-  } else if (_is_huge_sub_record || (len > buffer_size() - position())) {
-    // This object will not fit in completely or the last sub-record was huge.
-    // Finish the current segment and try again.
-    finish_dump_segment();
-    start_sub_record(tag, len);
-
-    return;
-  }
-
-  debug_only(_sub_record_left = len);
-  debug_only(_sub_record_ended = false);
-
-  write_u1(tag);
-}
-
-void AbstractDumpWriter::end_sub_record() {
-  assert(_in_dump_segment, "must be in dump segment");
-  assert(_sub_record_left == 0, "sub-record not written completely");
-  assert(!_sub_record_ended, "Must not have ended yet");
-  debug_only(_sub_record_ended = true);
-}
-
-// Supports I/O operations for a dump
-
-class DumpWriter : public AbstractDumpWriter {
-private:
-  FileWriter* _writer;
-  AbstractCompressor* _compressor;
-  size_t _bytes_written;
-  char* _error;
-  // Compression support
-  char* _out_buffer;
-  size_t _out_size;
-  size_t _out_pos;
-  char* _tmp_buffer;
-  size_t _tmp_size;
-
-private:
-  void do_compress();
-
-public:
-  DumpWriter(const char* path, bool overwrite, AbstractCompressor* compressor);
-  ~DumpWriter();
-  julong bytes_written() const override        { return (julong) _bytes_written; }
-  void set_bytes_written(julong bytes_written) { _bytes_written = bytes_written; }
-  char const* error() const override           { return _error; }
-  void set_error(const char* error)            { _error = (char*)error; }
-  bool has_error() const                       { return _error != nullptr; }
-  const char* get_file_path() const            { return _writer->get_file_path(); }
-  AbstractCompressor* compressor()             { return _compressor; }
-  void set_compressor(AbstractCompressor* p)   { _compressor = p; }
-  bool is_overwrite() const                    { return _writer->is_overwrite(); }
-  int get_fd() const                           { return _writer->get_fd(); }
-
-  void flush() override;
-};
-
-DumpWriter::DumpWriter(const char* path, bool overwrite, AbstractCompressor* compressor) :
-  AbstractDumpWriter(),
-  _writer(new (std::nothrow) FileWriter(path, overwrite)),
-  _compressor(compressor),
-  _bytes_written(0),
-  _error(nullptr),
-  _out_buffer(nullptr),
-  _out_size(0),
-  _out_pos(0),
-  _tmp_buffer(nullptr),
-  _tmp_size(0) {
-  _error = (char*)_writer->open_writer();
-  if (_error == nullptr) {
-    _buffer = (char*)os::malloc(io_buffer_max_size, mtInternal);
-    if (compressor != nullptr) {
-      _error = (char*)_compressor->init(io_buffer_max_size, &_out_size, &_tmp_size);
-      if (_error == nullptr) {
-        if (_out_size > 0) {
-          _out_buffer = (char*)os::malloc(_out_size, mtInternal);
-        }
-        if (_tmp_size > 0) {
-          _tmp_buffer = (char*)os::malloc(_tmp_size, mtInternal);
-        }
-      }
-    }
-  }
-  // initialize internal buffer
-  _pos = 0;
-  _size = io_buffer_max_size;
-}
-
-DumpWriter::~DumpWriter(){
-  if (_buffer != nullptr) {
-    os::free(_buffer);
-  }
-  if (_out_buffer != nullptr) {
-    os::free(_out_buffer);
-  }
-  if (_tmp_buffer != nullptr) {
-    os::free(_tmp_buffer);
-  }
-  if (_writer != NULL) {
-    delete _writer;
-  }
-  _bytes_written = -1;
-}
-
-// flush any buffered bytes to the file
-void DumpWriter::flush() {
-  if (_pos <= 0) {
-    return;
-  }
-  if (has_error()) {
-    _pos = 0;
-    return;
-  }
-  char* result = nullptr;
-  if (_compressor == nullptr) {
-    result = (char*)_writer->write_buf(_buffer, _pos);
-    _bytes_written += _pos;
-  } else {
-    do_compress();
-    if (!has_error()) {
-      result = (char*)_writer->write_buf(_out_buffer, _out_pos);
-      _bytes_written += _out_pos;
-    }
-  }
-  _pos = 0; // reset pos to make internal buffer available
-
-  if (result != nullptr) {
-    set_error(result);
-  }
-}
-
-void DumpWriter::do_compress() {
-  const char* msg = _compressor->compress(_buffer, _pos, _out_buffer, _out_size,
-                                          _tmp_buffer, _tmp_size, &_out_pos);
-
-  if (msg != nullptr) {
-    set_error(msg);
-  }
-}
 
 class DumperClassCacheTable;
 class DumperClassCacheTableEntry;
